+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### June
 
+- Update VS Code Extensions
+- Use Hyprland [scroll-overview](https://github.com/yayuuu/hyprland-scroll-overview)
 - Sandbox Firefox using [NixPak](https://github.com/nixpak/nixpak)
 - Add Zed Editor
 - Rename `nixos` script to `os` and adopt [`nh`](https://github.com/nix-community/nh)